Fix tests being run in parallel, failing SDL init
The cargo benchmarks also double as tests and so cause the `cargo test` command to fail, as they were not meant to be run in the test harness: the benchmark harness runs each test in sequence as to avoid resource contention, but tests are run in parallel by default.

Multiple SDL contexts can't coexist, so while sequential tests drop the context in time for the next test, parallel tests conflict with eachother for the global context slot and panic/fail.

This is not the best fix but is held up on https://github.com/rust-lang/cargo/issues/8430